requirements.txt
torch
diffusers
transformers
accelerate
peft
safetensors
pillow
runpod
fastapi
pydantic
uvicorn
huggingface-hub

